HTML to DMW XML siteimport
All XML config is set in 'config.json'. How to is coming.
but..
If number of 'slotfiles' don't maches the number settings in 'template',
'DMWslotsname', 'rank', 'context_id', 'config_id' and/or
'start_end_date' it's set to take the first for the setting to rest of
the settings. :) test it and see for yourself. Enjoy